+++ domains/client/domain-executor/src/lib.rs
@@ -87,21 +87,21 @@ mod domain_block_processor;
 mod domain_bundle_producer;
 mod domain_bundle_proposer;
 mod domain_worker;
+mod executor;
 mod fraud_proof;
 mod gossip_message_validator;
 mod parent_chain;
 mod sortition;
 mod system_bundle_processor;
 mod system_domain_worker;
-mod system_executor;
 mod system_gossip_message_validator;
 // TODO: Unlock once domain test infra is workable again.
 // #[cfg(test)]
 // mod tests;
 mod utils;
 
+pub use self::executor::Executor;
 pub use self::parent_chain::{CoreDomainParentChain, SystemDomainParentChain};
-pub use self::system_executor::Executor as SystemExecutor;
 pub use self::system_gossip_message_validator::SystemGossipMessageValidator;
 pub use self::utils::{DomainBlockImportNotification, DomainImportNotifications};
 use crate::utils::BlockInfo;
